UnliftIO.IO

Description

Unlifted System.IO.

Since: 0.1.0.0

Documentation

data IOMode

See openFile

Constructors

ReadMode

WriteMode

AppendMode

ReadWriteMode

data Handle

Haskell defines operations to read and write characters from and to files, represented by values of type Handle. Each value of this type is a handle: a record used by the Haskell run-time system to manage I/O with file system objects. A handle has at least the following properties:

whether it manages input or output or both;

whether it is open, closed or semi-closed;

whether the object is seekable;

whether buffering is disabled, or enabled on a line or block basis;

a buffer (whose length may be zero).

Most handles will also have a current I/O position indicating where the next input or output operation will occur. A handle is readable if it manages only input or both input and output; likewise, it is writable if it manages only output or both input and output. A handle is open when first allocated. Once it is closed it can no longer be used for either input or output, though an implementation cannot re-use its storage while references remain to it. Handles are in the Show and Eq classes. The string produced by showing a handle is system dependent; it should include enough information to identify the handle for debugging. A handle is equal according to == only to itself; no attempt is made to compare the internal state of different handles for equality.

stdin :: Handle

stdout :: Handle

stderr :: Handle

withFile :: MonadUnliftIO m => FilePath -> IOMode -> (Handle -> m a) -> m a

Unlifted version of withFile.

Since: 0.1.0.0

withBinaryFile :: MonadUnliftIO m => FilePath -> IOMode -> (Handle -> m a) -> m a

Unlifted version of withBinaryFile.

Since: 0.1.0.0

openFile :: MonadIO m => FilePath -> IOMode -> m Handle

Lifted version of openFile

Since: 0.2.20

hClose :: MonadIO m => Handle -> m ()

Lifted version of hClose

Since: 0.2.1.0

hFlush :: MonadIO m => Handle -> m ()

Lifted version of hFlush

Since: 0.2.1.0

hFileSize :: MonadIO m => Handle -> m Integer

Lifted version of hFileSize

Since: 0.2.1.0

hSetFileSize :: MonadIO m => Handle -> Integer -> m ()

Lifted version of hSetFileSize

Since: 0.2.1.0

hIsEOF :: MonadIO m => Handle -> m Bool

Lifted version of hIsEOF

Since: 0.2.1.0

data BufferMode

Three kinds of buffering are supported: line-buffering, block-buffering or no-buffering. These modes have the following effects. For output, items are written out, or flushed, from the internal buffer according to the buffer mode:

line-buffering: the entire output buffer is flushed whenever a newline is output, the buffer overflows, a hFlush is issued, or the handle is closed.

block-buffering: the entire buffer is written out whenever it overflows, a hFlush is issued, or the handle is closed.

no-buffering: output is written immediately, and never stored in the buffer.

An implementation is free to flush the buffer more frequently, but not less frequently, than specified above. The output buffer is emptied as soon as it has been written out.

Similarly, input occurs according to the buffer mode for the handle:

line-buffering: when the buffer for the handle is not empty, the next item is obtained from the buffer; otherwise, when the buffer is empty, characters up to and including the next newline character are read into the buffer. No characters are available until the newline character is available or the buffer is full.

block-buffering: when the buffer for the handle becomes empty, the next block of data is read into the buffer.

no-buffering: the next input item is read and returned. The hLookAhead operation implies that even a no-buffered handle may require a one-character buffer.

The default buffering mode when a handle is opened is implementation-dependent and may depend on the file system object which is attached to that handle. For most implementations, physical files will normally be block-buffered and terminals will normally be line-buffered.

Constructors

NoBuffering
buffering is disabled if possible.

LineBuffering
line-buffering should be enabled if possible.

BlockBuffering (Maybe Int)
block-buffering should be enabled if possible. The size of the buffer is n items if the argument is Just n and is otherwise implementation-dependent.

hSetBuffering :: MonadIO m => Handle -> BufferMode -> m ()

Lifted version of hSetBuffering

Since: 0.2.1.0

hGetBuffering :: MonadIO m => Handle -> m BufferMode

Lifted version of hGetBuffering

Since: 0.2.1.0

hSeek :: MonadIO m => Handle -> SeekMode -> Integer -> m ()

Lifted version of hSeek

Since: 0.2.1.0

data SeekMode

A mode that determines the effect of hSeek hdl mode i.

Constructors

AbsoluteSeek
the position of hdl is set to i.

RelativeSeek
the position of hdl is set to offset i from the current position.

SeekFromEnd
the position of hdl is set to offset i from the end of the file.

hTell :: MonadIO m => Handle -> m Integer

Lifted version of hTell

Since: 0.2.1.0

hIsOpen :: MonadIO m => Handle -> m Bool

Lifted version of hIsOpen

Since: 0.2.1.0

hIsClosed :: MonadIO m => Handle -> m Bool

Lifted version of hIsClosed

Since: 0.2.1.0

hIsReadable :: MonadIO m => Handle -> m Bool

Lifted version of hIsReadable

Since: 0.2.1.0

hIsWritable :: MonadIO m => Handle -> m Bool

Lifted version of hIsWritable

Since: 0.2.1.0

hIsSeekable :: MonadIO m => Handle -> m Bool

Lifted version of hIsSeekable

Since: 0.2.1.0

hIsTerminalDevice :: MonadIO m => Handle -> m Bool

Lifted version of hIsTerminalDevice

Since: 0.2.1.0

hSetEcho :: MonadIO m => Handle -> Bool -> m ()

Lifted version of hSetEcho

Since: 0.2.1.0

hGetEcho :: MonadIO m => Handle -> m Bool

Lifted version of hGetEcho

Since: 0.2.1.0

hWaitForInput :: MonadIO m => Handle -> Int -> m Bool

Lifted version of hWaitForInput

Since: 0.2.1.0

hReady :: MonadIO m => Handle -> m Bool

Lifted version of hReady

Since: 0.2.1.0

getMonotonicTime :: MonadIO m => m Double

Get the number of seconds which have passed since an arbitrary starting time, useful for calculating runtime in a program.

Since: 0.2.3.0