UnliftIO.MVar

Description

Unlifted Control.Concurrent.MVar.

Since: 0.1.0.0

Documentation

data MVar a :: * -> *

An MVar (pronounced "em-var") is a synchronising variable, used for communication between concurrent threads. It can be thought of as a a box, which may be empty or full.

newEmptyMVar :: MonadIO m => m (MVar a)

Lifted newEmptyMVar.

Since: 0.1.0.0

newMVar :: MonadIO m => a -> m (MVar a)

Lifted newMVar.

Since: 0.1.0.0

takeMVar :: MonadIO m => MVar a -> m a

Lifted takeMVar.

Since: 0.1.0.0

putMVar :: MonadIO m => MVar a -> a -> m ()

Lifted putMVar.

Since: 0.1.0.0

readMVar :: MonadIO m => MVar a -> m a

Lifted readMVar.

Since: 0.1.0.0

swapMVar :: MonadIO m => MVar a -> a -> m a

Lifted swapMVar.

Since: 0.1.0.0

tryTakeMVar :: MonadIO m => MVar a -> m (Maybe a)

Lifted tryTakeMVar.

Since: 0.1.0.0

tryPutMVar :: MonadIO m => MVar a -> a -> m Bool

Lifted tryPutMVar.

Since: 0.1.0.0

isEmptyMVar :: MonadIO m => MVar a -> m Bool

Lifted isEmptyMVar.

Since: 0.1.0.0

withMVar :: MonadUnliftIO m => MVar a -> (a -> m b) -> m b

Unlifted withMVar.

Since: 0.1.0.0

withMVarMasked :: MonadUnliftIO m => MVar a -> (a -> m b) -> m b

Unlifted withMVarMasked.

Since: 0.1.0.0

modifyMVar :: MonadUnliftIO m => MVar a -> (a -> m (a, b)) -> m b

Unlifted modifyMVar.

Since: 0.1.0.0

modifyMVar_ :: MonadUnliftIO m => MVar a -> (a -> m a) -> m ()

Unlifted modifyMVar_.

Since: 0.1.0.0

modifyMVarMasked :: MonadUnliftIO m => MVar a -> (a -> m (a, b)) -> m b

Unlifted modifyMVarMasked.

Since: 0.1.0.0

modifyMVarMasked_ :: MonadUnliftIO m => MVar a -> (a -> m a) -> m ()

Unlifted modifyMVarMasked_.

Since: 0.1.0.0

tryReadMVar :: MonadIO m => MVar a -> m (Maybe a)

Lifted tryReadMVar.

Since: 0.1.0.0

mkWeakMVar :: MonadUnliftIO m => MVar a -> m () -> m (Weak (MVar a))

Unlifted mkWeakMVar.

Since: 0.1.0.0